# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: goimg-datalayer API
  version: 2.0.0
  description: |
    A self-hosted image gallery backend inspired by Flickr and Chevereto.

    Features:
    - User management with JWT authentication
    - Image upload with automatic variant generation
    - Albums, tags, and social features (likes, comments)
    - Content moderation and abuse reporting
    - RESTful API with RFC 7807 error responses

    **Phase 2 Complete**: All Phase 2 features (Sprints 10-15) implemented.
    **Phase 3 Progress**: Sprints 16-20 complete (oEmbed, nested albums, custom variants, test coverage, trending tags, featured picks, groups/communities).
    **Current Sprint**: Sprint 23 in progress (test coverage improvement, documentation updates).
    Future features in backlog: video support (Sprint 21) and account tiers (Sprint 22).

  contact:
    name: goimg-datalayer API Team
    url: https://github.com/yegamble/goimg-datalayer
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.goimg.com/v1
    description: Production server (example)
security:
  - bearerAuth: []
...
paths:
...
  /images:
    post:
      tags: [images]
      summary: Upload image
      description: |
        Upload a new image with optional metadata.

        Processing:
        - ClamAV malware scan (required)
        - EXIF metadata extraction
        - EXIF stripping (GPS data removed by default)
        - Generate 4 variants: thumbnail (150px), small (320px), medium (800px), large (1600px)
        - Store original
        - Re-encode through libvips for security

        Rate limit: 50 uploads per hour per user
      operationId: uploadImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, GIF, WebP)
                title:
                  type: string
                  maxLength: 255
                  example: Sunset over mountains
                description:
                  type: string
                  maxLength: 2000
                  example: Beautiful sunset captured in Rocky Mountains
                visibility:
                  type: string
                  enum: [public, private, unlisted]
                  default: private
                album_id:
                  type: string
                  format: uuid
                  description: Add image to album (optional)
                tags:
                  type: string
                  description: Comma-separated tags (max 20)
                  example: sunset,mountains,nature,landscape
      responses:
        '201':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '413':
          description: File too large (max 10MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '415':
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '422':
          description: Malware detected or processing failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '429':
          description: Upload rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    get:
      tags: [images]
      summary: List images
      description: |
        List images with optional filtering and pagination.

        Returns only images the authenticated user has permission to view:
        - Own images (all visibility levels)
        - Public images from other users
      operationId: listImages
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: owner_id
          in: query
          description: Filter by image owner
          schema:
            type: string
            format: uuid
        - name: album_id
          in: query
          description: Filter by album
          schema:
            type: string
            format: uuid
        - name: visibility
          in: query
          description: Filter by visibility (only for own images)
          schema:
            type: string
            enum: [public, private, unlisted]
        - name: tags
          in: query
          description: Filter by tags (comma-separated, AND logic)
          schema:
            type: string
            example: sunset,nature
        - name: q
          in: query
          description: Search query (title, description)
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [created_at, view_count, like_count]
            default: created_at
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Images retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Image'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /images/{id}/variants:
    post:
      tags: [images]
      summary: Generate custom variant
      description: |
        Generate a custom image variant using user-defined parameters or a saved variant config.

        You can either:
        1. Provide a config_id to use a saved variant configuration
        2. Provide direct parameters (max_width, max_height, format, quality, crop_mode)

        Crop modes:
        - fit: Scale to fit within dimensions, preserving aspect ratio (default)
        - fill: Scale to fill dimensions, cropping excess
        - crop: Crop from center to exact dimensions

        Output formats:
        - jpeg: Standard JPEG
        - png: PNG with transparency support
        - webp: Modern WebP format (default, best compression)
        - avif: Next-gen AVIF format

        Only image owners can generate custom variants.
      operationId: generateCustomVariant
      parameters:
        - $ref: '#/components/parameters/ImageIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                config_id:
                  type: string
                  format: uuid
                  description: Use a saved variant config (alternative to direct parameters)
                name:
                  type: string
                  maxLength: 50
                  description: Custom name for the variant (used in storage key)
                  example: "social-square"
                max_width:
                  type: integer
                  minimum: 1
                  maximum: 8192
                  description: Maximum width in pixels (required if config_id not provided)
                  example: 500
                max_height:
                  type: integer
                  minimum: 1
                  maximum: 8192
                  description: Maximum height in pixels (required if config_id not provided)
                  example: 500
                format:
                  type: string
                  enum: [jpeg, png, webp, avif]
                  default: webp
                  example: webp
                quality:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 85
                  example: 85
                crop_mode:
                  type: string
                  enum: [fit, fill, crop]
                  default: fit
                  example: fit
            examples:
              using_config:
                summary: Using a saved variant config
                value:
                  config_id: "123e4567-e89b-12d3-a456-426614174000"
              direct_params:
                summary: Direct parameters
                value:
                  name: "instagram-square"
                  max_width: 1080
                  max_height: 1080
                  format: "jpeg"
                  quality: 90
                  crop_mode: "fill"
      responses:
        '201':
          description: Custom variant generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomVariantResult'
        '400':
          description: Invalid parameters or missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Forbidden - only image owner can generate variants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Image not found or variant config not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          description: Processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /images/search:
    get:
      tags: [images]
      summary: Search images
      description: |
        Full-text search across image titles and descriptions.
        Supports filtering by tags, owner, and visibility.
      operationId: searchImages
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Search query (searches title and description)
          schema:
            type: string
            minLength: 1
            example: sunset mountains
        - name: tags
          in: query
          description: Filter by tag name
          schema:
            type: string
            example: nature
        - name: owner_id
          in: query
          description: Filter by owner ID
          schema:
            type: string
            format: uuid
        - name: visibility
          in: query
          description: Filter by visibility (only for own images)
          schema:
            type: string
            enum: [public, private, unlisted]
        - name: sort_by
          in: query
          description: Sort field
          schema:
            type: string
            enum: [relevance, created_at, view_count, like_count]
            default: relevance
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Image'
        '400':
          description: Invalid search query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  # ==================== Albums ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer authentication. Include the access token in the Authorization header:

        ```
        Authorization: Bearer <access_token>
        ```

        Access tokens expire after 15 minutes. Use the refresh endpoint to obtain a new token.
  parameters:
...
    ImageIdParam:
      name: id
      in: path
      required: true
      description: Image ID
      schema:
        type: string
        format: uuid
      example: 123e4567-e89b-12d3-a456-426614174000
...
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    PerPageParam:
      name: per_page
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
  schemas:
...
    Image:
      type: object
      required: [id, owner_id, owner, visibility, mime_type, width, height, variants, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        owner_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
        owner:
          type: object
          required: [id, username]
          properties:
            id:
              type: string
              format: uuid
            username:
              type: string
            display_name:
              type: string
              nullable: true
          example:
            id: 123e4567-e89b-12d3-a456-426614174001
            username: johndoe
            display_name: John Doe
        title:
          type: string
          nullable: true
          example: Sunset over mountains
        description:
          type: string
          nullable: true
          example: Beautiful sunset captured in Rocky Mountains
        visibility:
          type: string
          enum: [public, private, unlisted]
          example: public
        mime_type:
          type: string
          example: image/jpeg
        file_size:
          type: integer
          description: File size in bytes
          example: 2048576
        width:
          type: integer
          description: Image width in pixels
          example: 4000
        height:
          type: integer
          description: Image height in pixels
          example: 3000
        variants:
          type: object
          properties:
            thumbnail:
              $ref: '#/components/schemas/ImageVariant'
            small:
              $ref: '#/components/schemas/ImageVariant'
            medium:
              $ref: '#/components/schemas/ImageVariant'
            large:
              $ref: '#/components/schemas/ImageVariant'
            original:
              $ref: '#/components/schemas/ImageVariant'
        tags:
          type: array
          items:
            type: string
          example: [sunset, mountains, nature, landscape]
        view_count:
          type: integer
          example: 150
        like_count:
          type: integer
          example: 42
        comment_count:
          type: integer
          example: 8
        created_at:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
        updated_at:
          type: string
          format: date-time
          example: 2024-01-15T11:00:00Z
    ImageVariant:
      type: object
      required: [url, width, height]
      properties:
        url:
          type: string
          format: uri
          example: https://cdn.goimg.com/images/123e4567/medium.jpg
        width:
          type: integer
          example: 800
        height:
          type: integer
          example: 600

    # ==================== IPFS Schemas ====================
...
    CustomVariantResult:
      type: object
      description: Result of custom variant generation
      required: [variant_key, variant_url, width, height, format, file_size, content_type]
      properties:
        variant_key:
          type: string
          description: Storage key for the generated variant
          example: "images/123e4567-e89b-12d3-a456-426614174000/custom/social-square.webp"
        variant_url:
          type: string
          format: uri
          description: URL to access the generated variant
          example: "https://cdn.example.com/images/123e4567-e89b-12d3-a456-426614174000/custom/social-square.webp"
        width:
          type: integer
          description: Actual width of the generated variant
          example: 500
        height:
          type: integer
          description: Actual height of the generated variant
          example: 500
        format:
          type: string
          description: Output format of the variant
          example: "webp"
        file_size:
          type: integer
          format: int64
          description: File size in bytes
          example: 45678
        content_type:
          type: string
          description: MIME type of the variant
          example: "image/webp"
...
    PaginatedResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items: {}
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      required: [total, page, per_page, total_pages]
      properties:
        total:
          type: integer
          description: Total number of items
          example: 150
        page:
          type: integer
          description: Current page number
          example: 1
        per_page:
          type: integer
          description: Items per page
          example: 20
        total_pages:
          type: integer
          description: Total number of pages
          example: 8

    # ==================== Error Responses (RFC 7807) ====================
    ProblemDetail:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
          example: https://api.goimg.com/problems/validation-error
        title:
          type: string
          description: A short, human-readable summary of the problem type
          example: Validation Error
        status:
          type: integer
          description: The HTTP status code
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
          example: Request validation failed
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence
          example: /api/v1/images
        traceId:
          type: string
          description: A unique identifier for tracing the request
          example: abc123-def456
        errors:
          type: array
          description: Detailed validation errors (for 400 responses)
          items:
            type: object
            properties:
              field:
                type: string
                example: title
              message:
                type: string
                example: must not exceed 255 characters

    # ==================== Session ====================
...